Rename library from snowplow_tracker to aio_snowplow_tracker
This allows this library to be installed and used alongside the official snowplow_tracker library.
