+++ .github/workflows/ci.yml
@@ -51,7 +51,7 @@ jobs:
 
     - name: Tests
       run: |
-        pytest --cov=snowplow_tracker --cov-report=xml
+        pytest --cov=aio_snowplow_tracker --cov-report=xml
 
     - name: Coveralls
       uses: AndreMiras/coveralls-python-action@develop
