--- snowplow_tracker/__init__.py
@@ -1,12 +1,0 @@
-from snowplow_tracker._version import __version__
-from snowplow_tracker.subject import Subject
-from snowplow_tracker.emitters import logger, Emitter
-from snowplow_tracker.self_describing_json import SelfDescribingJson
-from snowplow_tracker.tracker import Tracker
-from snowplow_tracker.contracts import disable_contracts, enable_contracts
-
-# celery extra
-from .celery import CeleryEmitter
-
-# redis extra
-from .redis import RedisEmitter, RedisWorker
